Handle DELETE_SELF event to prevent interruption
Some programs does not directly write into the monitored file but, instead,
write into a temporary file, delete the original one before renaming the
temporary file to the monitored file name.
This workflow has not been handled correctly in file_backup.
When the file is deleted, inotifywait emits the DELETE_SELF event and
stops to watch the file by design.
This commit defines the recursive function file_monitor that monitors
the DELETE_SELF event and kills and re-execute inotifywait when that
event is emitted.